individual Employee Sentiment Analysis
It is made for understanding individual employee's sentiments that showcases their overall time at the organization as positive or negative. For this purpose, a sample was taken as a dataset in this file and was filtered on the bases of the questionnaires asked by the Human Resource professional to analyze as to what positive opinion does the employee has towards the exiting organization. Showcased through pie chart how positive the feedback is given by the employee for the company.